// site/src/lib/database.d.ts
export interface Database {
// ...
  public: {
    Tables: {
      guilds: {
        Row: {
          created_at: string
          guild_id: number
        }
        Insert: {
          created_at?: string
          guild_id: number
        }
        Update: {
          created_at?: string
          guild_id?: number
        }
      }
      provider_data: {
        Row: {
          access_token: string
          created_at: string
          expires_at: string
          id: string
          refresh_token: string
          scopes: string
          token_type: string
        }
        Insert: {
          access_token: string
          created_at?: string
          expires_at: string
          id: string
          refresh_token: string
          scopes: string
          token_type: string
        }
        Update: {
          access_token?: string
          created_at?: string
          expires_at?: string
          id?: string
          refresh_token?: string
          scopes?: string
          token_type?: string
        }
      }
      puns: {
        Row: {
          created_at: string
          guild_id: number | null
          id: number
          pun: string
        }
        Insert: {
          created_at?: string
          guild_id?: number | null
          id?: number
          pun: string
        }
        Update: {
          created_at?: string
          guild_id?: number | null
          id?: number
          pun?: string
        }
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}